Update scripts structure
